Removed obsolete method LogAssemblyVersion from InternalLogger
